Rework tests to extract out the user from ping tests
